refactor: api호출 필터링 로직 분리
